chore: move proxy message to package
